requirements.txt
PyPDF2>=3.0.0
pdfminer.six>=20221105
ReportLab>=3.6.12
pillowcase>=2.0.0
Pillow>=9.4.0
python-docx>=0.8.11
markdown>=3.4.3
requests>=2.25.0
flask>=2.0.0
werkzeug>=2.0.0
pymupdf>=1.21.0  # PyMuPDF for better PDF handling and direct manipulation
pdfrw>=0.4       # Alternative PDF manipulation library
pikepdf>=5.0.0   # Another option for direct PDF manipulation
pytesseract>=0.3.10  # For OCR capabilities
pdf2image>=1.16.0  # For converting PDF to images for OCR
numpy>=1.21.0  # For array operations
opencv-python>=4.5.0  # Optional but recommended for better image preprocessing
scikit-learn>=1.0.0  # For ML-based section classification (optional)
pillow>=9.4.0  # For image handling
pymupdf>=1.21.0  # For better PDF handling (fitz)
